+++ src/moria/index.ts
@@ -1,0 +1,1 @@
+export { default as MoriaV0 } from './v0/index.js';